# openapi.yaml
openapi: 3.0.3
info:
  title: Lymchic Store API
  version: 1.0.0
  description: API RESTful para la tienda Lymchic Store
servers:
  - url: http://localhost:3000/api
paths:
...
  /categories:
    get:
      summary: Listar categorías
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [name, description, id, image, featured] }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: Lista de categorías
          content:
            application/json:
              example:
                data:
                  - id: "a1b2c3d4-5678-1234-9abc-1234567890ab"
                    name: Ropa
                    description: Ropa de moda
                    image: https://img.com/catropa.png
                    featured: true
                pagination:
                  total: 1
                  page: 1
                  limit: 10
                  totalPages: 1
  /categories/{id}:
    get:
      summary: Obtener categoría por ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Categoría encontrada
          content:
            application/json:
              example:
                id: "a1b2c3d4-5678-1234-9abc-1234567890ab"
                name: Ropa
                description: Ropa de moda
                image: https://img.com/catropa.png
                featured: true
        '400':
          description: ID inválido
          content:
            application/json:
              example:
                message: Invalid category ID
        '404':
          description: Categoría no encontrada
          content:
            application/json:
              example:
                message: Category not found
  /categories/admin:
    post:
      summary: Crear categoría (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                image: { type: string }
                featured: { type: boolean }
              required: [name, description, image, featured]
            example:
              name: Ropa
              description: Ropa de moda
              image: https://img.com/catropa.png
              featured: true
      responses:
        '201':
          description: Categoría creada
          content:
            application/json:
              example:
                id: "a1b2c3d4-5678-1234-9abc-1234567890ab"
                name: Ropa
                description: Ropa de moda
                image: https://img.com/catropa.png
                featured: true
        '400':
          description: Error de validación
          content:
            application/json:
              example:
                message: Validation Error
                errors:
                  - path: [name]
                    message: Required
  /orders:
    post:
      summary: Crear orden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                delivery_method: { type: string }
                order_details:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id: { type: string, format: uuid }
                      quantity: { type: integer }
                      unit_price: { type: number }
                    required: [product_id, quantity, unit_price]
              required: [delivery_method, order_details]
            example:
              delivery_method: STANDARD
              order_details:
                - product_id: "b1e2c3d4-5678-1234-9abc-1234567890ab"
                  quantity: 2
                  unit_price: 25.5
      responses:
        '201':
          description: Orden creada
          content:
            application/json:
              example:
                order_id: 1
                order_number: ORD-123456
                user_id: 1
                order_date: "2025-07-05T12:00:00.000Z"
                order_status_id: 1
                delivery_method: STANDARD
        '400':
          description: Error de validación
          content:
            application/json:
              example:
                message: Validation Error
                errors:
                  - path: [order_details]
                    message: Required
...
  /order-statuses/admin:
    post:
      summary: Crear estado de orden (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status_name: { type: string }
              required: [status_name]
            example:
              status_name: PENDING
      responses:
        '201':
          description: Estado creado
          content:
            application/json:
              example:
                order_status_id: 1
                status_name: PENDING
        '400':
          description: Error de validación
          content:
            application/json:
              example:
                message: Validation Error
                errors:
                  - path: [status_name]
                    message: Required